Split GeometryInfo expressions out
