+++ src/engine/sparqlExpressions/CMakeLists.txt
@@ -22,6 +22,7 @@ add_library(sparqlExpressions
         CountStarExpression.cpp
         PrefilterExpressionIndex.cpp
         GeoExpression.cpp
+        GeometryInfoExpression.cpp
         BlankNodeExpression.cpp
         GroupConcatExpression.cpp)
